chore: move header to folder for component abstraction
