+++ x-pack/solutions/search/plugins/search_homepage/public/components/header/index.tsx
@@ -18,7 +18,8 @@ import {
   EuiLink,
 } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
-import { useAssetBasePath } from '../hooks/use_asset_base_path';
+
+import { useAssetBasePath } from '../../hooks/use_asset_base_path';
 
 export const SearchHomepageHeader: React.FC = () => {
   const { euiTheme, colorMode } = useEuiTheme();
